+++ src/collections.rs
@@ -148,6 +148,7 @@ macro_rules! btset {
 ///
 /// assert_eq!(deque, deque2);
 /// # }
+/// ```
 ///
 /// [`VecDeque`]: https://doc.rust-lang.org/std/collections/struct.VecDeque.html
 #[macro_export]
@@ -280,6 +281,7 @@ macro_rules! flkl {
 ///     println!("{}", i);
 /// }
 /// # }
+/// ```
 ///
 /// [`BinaryHeap`]: https://doc.rust-lang.org/std/collections/struct.BinaryHeap.html
 #[macro_export]