+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changes
 
-## Unreleased ~ master
+## Unreleased ~ master ~ 2.0-dev
+ - Remove haskell-like and python-like comprehension syntax
+ - Re-implement comprehensions based on the `c!` macro
+ - Add a comprehension macro for lazy iterators: `c!`
 
 ## 1.2.0
  - Add a collection macro for `BinaryHeap`: `bheap!`